Finish CRUD product combo

+++ LowLand/View/ViewModel/ProductsViewModel.cs
@@ -18,9 +18,30 @@ namespace LowLand.View.ViewModel
 
         public ResponseCode RemoveProduct(int productId)
         {
+            bool isChildProduct = _dao.ComboItems.GetAll()
+                .Any(item => item.Product.Id == productId);
+
+            if (isChildProduct)
+            {
+                return ResponseCode.ItemHaveDependency;
+            }
+
             var product = Products.FirstOrDefault(p => p.Id == productId);
             if (product != null)
             {
+                // If the product is a combo, remove all child products
+                if (product is ComboProduct)
+                {
+                    var childProducts = _dao.ComboItems.GetAll()
+                        .Where(item => item.ComboId == productId)
+                        .ToList();
+
+                    foreach (var item in childProducts)
+                    {
+                        _dao.ComboItems.DeleteById(item.ItemId.ToString());
+                    }
+                }
+
                 int result = _dao.Products.DeleteById(productId.ToString());
 
                 if (result != -1)
@@ -31,13 +52,11 @@ namespace LowLand.View.ViewModel
                     {
                         return ResponseCode.Error;
                     }
-
-                    return ResponseCode.Success;
                 }
-            }
 
+                return ResponseCode.Success;
+            }
             return ResponseCode.NotFound;
         }
-
     }
 }
